requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
async-timeout==4.0.2
attrs==22.1.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
discord.py==2.1.0
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
Flask==3.1.0
frozenlist==1.3.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.13.0
MarkupSafe==3.0.2
multidict==6.0.3
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.5.2
python-dotenv==1.2.1
requests==2.33.1
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.3
yarl==1.8.2
